refactor(remote): Fix import
