+++ zio-flow/shared/src/test/scala/zio/flow/remote/RemoteDurationSpec.scala
@@ -1,4 +1,4 @@
-package zio.flow
+package zio.flow.remote
 
 import zio.flow.utils.RemoteAssertionSyntax.RemoteAssertionOps
 import zio.test._
