+++ zio-flow/shared/src/test/scala/zio/flow/remote/RemoteEitherSpec.scala
@@ -1,4 +1,4 @@
-package zio.flow
+package zio.flow.remote
 
 import zio.flow.utils.RemoteAssertionSyntax.RemoteAssertionOps
 import zio.random.Random
@@ -9,7 +9,7 @@ object RemoteEitherSpec extends DefaultRunnableSpec {
     testM("handleEither") {
       check(Gen.either(Gen.anyInt, Gen.boolean)) { either =>
         val expected = either.fold(_ * 2, if (_) 10 else 20)
-        val result   = Remote(either).handleEither(_ * 2, _.ifThenElse(10, 20))
+        val result = Remote(either).handleEither(_ * 2, _.ifThenElse(10, 20))
         result <-> expected
       }
     },
